To fix the cleanup of the processed data from lease table
Problem: the unprocessed changes count was not coming to zero, so leases table was not getting cleanup. Which was causing increase in size of lease table.
Fix : modify query of unprocessed changes count and update query of lease table cleanup\delete rows which are completed or expired
add few more logs to easy track issue in future